Add 'discovery' experiment directory
This experiment is useful for playing with service discovery and
health-checking.

Create instances of simple 'hello' service which uses netcat to output
'hello' on port 80 and then shuts down. Discover running instances of
the services using Consul DNS and read the output using netcat.

Includes a worked example of the experiment.