leader: join cluster and watch at start, not after
'consul agent' supports a '-join' parameter so it's not necessary for us
to have a separate 'consul join' invocation.

This also removes the need for the retry loop on 'consul join'.

Additionally, we can now make the hello/leader watch part of the
configuration as the cluster will be joined on startup. Previously they
would fail if part of configuration as the cluster wouldn't be joined
until later.